Added managment service for transaction types, started to reorganize the code to a more unified standard. Lots of cool stuff in this one

+++ app/Http/Controllers/TransactionTypeController.php
@@ -2,17 +2,17 @@
 
 namespace App\Http\Controllers;
 
+use App\Http\Requests;
+use App\Models\Permission;
 use App\Models\TransactionType;
 use Illuminate\Http\Request;
-use App\Http\Requests;
-use App\Http\Controllers\Controller;
 use Laracasts\Flash\Flash;
 
 class TransactionTypeController extends Controller
 {
 
-    public function __construct() {
-        $this->middleware('admin');
+    public function __construct()
+    {
     }
 
     /**
@@ -23,7 +23,8 @@ class TransactionTypeController extends Controller
     public function index()
     {
         $types = TransactionType::All();
-        return view('transtype.index')->withTypes($types);
+        $permissions = Permission::All();
+        return view('transtype.index')->withTypes($types)->withPermissions($permissions);
     }
 
     /**
@@ -39,59 +40,74 @@ class TransactionTypeController extends Controller
     /**
      * Store a newly created resource in storage.
      *
-     * @param  \Illuminate\Http\Request  $request
+     * @param  \Illuminate\Http\Request $request
      * @return \Illuminate\Http\Response
      */
     public function store(Request $request)
     {
-        //
+        $transactionType = TransactionType::create($request->all());
     }
 
     /**
      * Display the specified resource.
      *
-     * @param  int  $id
+     * @param  int $id
      * @return \Illuminate\Http\Response
      */
-    public function show($id)
+    public function show($type)
     {
-        //
+        return $this->edit($type);
     }
 
     /**
      * Show the form for editing the specified resource.
      *
-     * @param  int  $id
+     * @param  int $id
      * @return \Illuminate\Http\Response
      */
-    public function edit($id)
+    public function edit($type)
     {
-        //
+
+        $permissions = Permission::All();
+
+        return view('transtype.edit')->withType($type)->withPermissions($permissions);
     }
 
     /**
      * Update the specified resource in storage.
      *
-     * @param  \Illuminate\Http\Request  $request
-     * @param  int  $id
+     * @param  \Illuminate\Http\Request $request
+     * @param  int $id
      * @return \Illuminate\Http\Response
      */
-    public function update(Request $request, $id)
+    public function update(Request $request, $type)
     {
-        //
+        if ($type->locked) {
+            Flash::error('Sorry, this is not editable.');
+            return redirect()->back();
+        }
+
+        $type->update($request->all());
+        Flash::success('You have updated the ' . $type->name . ' transaction type');
+        return redirect(url('transtype/' . $type->id));
     }
 
     /**
      * Remove the specified resource from storage.
      *
-     * @param  int  $id
+     * @param  int $id
      * @return \Illuminate\Http\Response
      */
     public function destroy($id)
     {
-        Flash::error('You deleted '.$transType->name);
         $transType = TransactionType::find($id);
-        $transType->delete();
+        if ($transType->locked) {
+            Flash::error('Sorry, this permission is essential to this application.');
+        } else {
+            $transType->delete();
+            Flash::success('You deleted ' . $transType->name);
+        }
+
         return redirect()->back();
     }
 }
